refactor: relocate safety_manifest from command_contract into cli_surface
command_contract/safety_manifest.rs builds the command safety manifest entirely
from cli_surface types (CommandSurface, CommandSafetyManifest, ...) and is used
only by cli_runtime and commands -- never by core. cli_surface already re-exports
its functions as the public interface, so command_contract was the wrong home.

Convert cli_surface.rs into a directory module (cli_surface/mod.rs) and move
safety_manifest.rs in as cli_surface/safety_manifest.rs. cli_surface now owns and
re-exports it directly; the safety_manifest -> command_contract dependency
(registered_command, CommandJsonFamily) keeps the correct cli_surface ->
command_contract direction. Fixed the include_str! path for the commands index
that shifted one directory deeper when cli_surface became a module dir.

This removes safety_manifest's cli_surface coupling from command_contract,
leaving only the descriptors dispatch macro (definition-only).

cargo check --lib and --lib --tests pass.
