+++ src/command_contract.rs
@@ -23,7 +23,6 @@ mod lab;
 mod output;
 mod public_variants;
 mod registry;
-pub mod safety_manifest;
 mod spec;
 
 pub use crate::core::artifact_ref::{
